Rename to app_settings for contextual sense
