+++ app/app_settings/models.py
