+++ app/app_settings/tests/tests_views.py
