+++ app/core/urls.py
@@ -6,7 +6,7 @@ from django.contrib import admin
 from django.urls import path
 from django.conf.urls import url
 from app.index.views import IndexView
-from app.settings_app.views import SettingsView, SettingsSave
+from app.app_settings.views import SettingsView, SettingsSave
 
 urlpatterns = [
     path('admin/settings/save', SettingsSave, name='settings_save'),
