--- app/settings_app/__init__.py
@@ -1,8 +1,0 @@
-from django.apps import AppConfig
-
-class SettingsAppConfig(AppConfig):
-    name = 'app.settings_app'
-    label = 'settings_app'
-    verbose_name = 'Settings App'
-
-default_app_config = 'app.settings_app.SettingsAppConfig'
